refactor test class for more DRY-ness
